+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 - Unreleased
     - Added documentation site
     - Added Regression and Classification Loss interfaces
+    - Robust Z Score is now a Ranking anomaly detector
     - LODA now defaults to auto detect bin count
     - Removed tolerance param from Gradient Boost and AdaBoost
     - Screen logger timestamp format now configurable
     - Dropped Persistable contract between SVM-based learners
     - Random Forest feature importances now serial
+    - Removed Robust Z Score tolerance parameter
 
 - 0.0.12-beta
     - Added AdaMax neural network Optimizer